refactor(tunnel): remove duplicated kube helpers
